+++ tools/gulptasks/lib/test.js
@@ -16,21 +16,6 @@
  * */
 
 /**
- * Fails if the product is not affected in the files staged for commit.
- *
- * @throws Will throw an error if the product is not affected.
- *
- * @param {string} product
- *        The product name that should be checked.
- */
-function checkProduct(product) {
-    const products = getProducts();
-    if (products.indexOf(product) === -1) {
-        throw new Error(`${product} is not affected`);
-    }
-}
-
-/**
  * Returns list of products affected by modified files staged for commit.
  *
  * @param {Boolean} logPaths
@@ -92,6 +77,122 @@ function getProducts(logPaths) {
     return affectedProducts;
 }
 
+/**
+ * Fails if the product is not affected in the files staged for commit.
+ *
+ * @throws Will throw an error if the product is not affected.
+ *
+ * @param {string} product
+ *        The product name that should be checked.
+ */
+function checkProduct(product) {
+    const products = getProducts();
+    if (products.indexOf(product) === -1) {
+        throw new Error(`${product} is not affected`);
+    }
+}
+
+/**
+ * Checks if tests should run
+ * @param {{ configFile: string, codeDirectory: string, jsDirectory: string, testsDirectory: string }}} config
+ * Configuration
+ *
+ * @return {Promise<boolean>}
+ * True if outdated
+ */
+async function shouldRun({
+    configFile,
+    codeDirectory,
+    jsDirectory,
+    testsDirectory
+}) {
+
+    const fs = require('fs');
+    const fsLib = require('./fs');
+    const logLib = require('./log');
+    const stringLib = require('./string');
+
+    let configuration = {
+        latestCodeHash: '',
+        latestJsHash: '',
+        latestTestsHash: ''
+    };
+
+    if (fs.existsSync(configFile)) {
+        configuration = JSON.parse(
+            fs.readFileSync(configFile).toString()
+        );
+    }
+
+    const latestCodeHash = fsLib.getDirectoryHash(
+        codeDirectory, true, stringLib.removeComments
+    );
+    const latestJsHash = fsLib.getDirectoryHash(
+        jsDirectory, true, stringLib.removeComments
+    );
+    const latestTestsHash = fsLib.getDirectoryHash(
+        testsDirectory, true, stringLib.removeComments
+    );
+
+    if (latestCodeHash === configuration.latestCodeHash &&
+        latestJsHash !== configuration.latestJsHash
+    ) {
+
+        throw new Error('Code out of sync');
+    }
+
+    if (latestCodeHash === configuration.latestCodeHash &&
+        latestTestsHash === configuration.latestTestsHash
+    ) {
+
+        logLib.success(
+            '✓ Source code and unit tests not have been modified' +
+            ' since the last successful test run.'
+        );
+
+        return false;
+    }
+
+    return true;
+}
+
+/**
+ * Saves test run information
+ * @param {{ configFile: string, codeDirectory: string, jsDirectory: string, testsDirectory: string }} config
+ * Configuration
+ *
+ * @return {void}
+ */
+function saveRun({
+    configFile,
+    codeDirectory,
+    jsDirectory,
+    testsDirectory
+}) {
+
+    const FS = require('fs');
+    const FSLib = require('./fs');
+    const StringLib = require('./string');
+
+    const latestCodeHash = FSLib.getDirectoryHash(
+        codeDirectory, true, StringLib.removeComments
+    );
+    const latestJsHash = FSLib.getDirectoryHash(
+        jsDirectory, true, StringLib.removeComments
+    );
+    const latestTestsHash = FSLib.getDirectoryHash(
+        testsDirectory, true, StringLib.removeComments
+    );
+
+    const configuration = {
+        latestCodeHash,
+        latestJsHash,
+        latestTestsHash
+    };
+
+    FS.writeFileSync(configFile, JSON.stringify(configuration));
+}
+
 /* *
  *
  *  Exports
@@ -100,5 +201,7 @@ function getProducts(logPaths) {
 
 module.exports = {
     checkProduct,
-    getProducts
+    getProducts,
+    shouldRun,
+    saveRun
 };
